Addressing Comments
