+++ src/workstation-gui/app/controllers/application_controller.rb
@@ -18,7 +18,6 @@ class ApplicationController < ActionController::Base
   # protect_from_forgery with: :null_session
   # before_action :authenticate_api_requests!
 
-  include ApplicationHelper
 
   def authenticate_api_requests!
     header = request.headers['Authorization']
